Use armored, normal asc format

+++ bin/setup_keys.sh
@@ -4,7 +4,7 @@ set -eu
 KEYS_DIR="/keys"
 GPG_DIR="$DRONE_DIR/.gnupg"
 SSH_DIR="$DRONE_DIR/.ssh"
-GPG_KEYS=("$KEYS_DIR/platform.pubring.asc" "$KEYS_DIR/platform.secring.asc")
+GPG_KEYS=("$KEYS_DIR/pubring.asc" "$KEYS_DIR/secring.asc")
 SSH_KEYS=("$KEYS_DIR/id_rsa" "$KEYS_DIR/id_rsa.pub")
 
 test -d "$GPG_DIR" || mkdir "$GPG_DIR"
@@ -13,7 +13,7 @@ test -d "$SSH_DIR" || mkdir "$SSH_DIR"
 if [[ -d $KEYS_DIR ]]; then
   for GPG_KEY in "${GPG_KEYS[@]}"; do
     GPG_FILENAME=$(basename "$GPG_KEY")
-    if [[ ! -e "$GPG_DIR/$GPG_FILENAME" ]]; then
+    if [[ -e "$GPG_KEY" && ! -e "$GPG_DIR/$GPG_FILENAME" ]]; then
       echo "Copying $GPG_KEY to $GPG_DIR."
       cp "$GPG_KEY" "$GPG_DIR"
     fi
@@ -21,7 +21,7 @@ if [[ -d $KEYS_DIR ]]; then
 
   for SSH_KEY in "${SSH_KEYS[@]}"; do
     SSH_FILENAME=$(basename "$SSH_KEY")
-    if [[ ! -e "$SSH_DIR/$SSH_FILENAME" ]]; then
+    if [[ -e "$SSH_KEY" && ! -e "$SSH_DIR/$SSH_FILENAME" ]]; then
       echo "Copying $SSH_KEY to $SSH_DIR."
       cp "$SSH_KEY" "$SSH_DIR"
       cp "$SSH_KEY" "$HOME/.ssh"
